Remove async PSO compilation
This was causing a race condition when 2 PSOs want to request the same shader. If requested at the same time, the shader compiler will compile it twice and one will stomp the other in the shader cache. I can't think of a good workaround for this at the moment. Very sad ;(